FirstProject/pyproject.toml
[project]
name = "firstproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.26",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.3",
    "python-dotenv>=1.1.1",
]
